feat: add sub-millisecond precision to response times
Change response_time_ms from int64 to float64 to support sub-millisecond
precision for very fast queries (cached, blocked, local records).

Changes:
- Update QueryLog.ResponseTimeMs from int64 to float64 in storage layer
- Update QueryResponse.ResponseTimeMs from int64 to float64 in API layer
- Update QueryData.ResponseTimeMs from int64 to float64 in UI handlers
- Change time.Since().Milliseconds() to time.Since().Seconds() * 1000
  for float64 precision in DNS query handler
- Add migration v3 to convert SQLite column from INTEGER to REAL
  while preserving existing data

Benefits:
- Cached queries now show ~0.03-0.05ms instead of 0ms
- Blocked queries show ~0.08-0.13ms instead of 0ms
- Better visibility into actual performance characteristics
- More informative metrics for sub-millisecond operations

Example output:
- Cached query: 0.037ms (37 microseconds)
- Blocked query: 0.086ms (86 microseconds)
- Upstream query: 4.298ms

Previously all fast queries showed 0ms which was not informative.

+++ pkg/api/responses.go
@@ -45,7 +45,7 @@ type QueryResponse struct {
 	Upstream       string                    `json:"upstream,omitempty"`
 	ID             int64                     `json:"id"`
 	ResponseCode   int                       `json:"response_code"`
-	ResponseTimeMs int64                     `json:"response_time_ms"`
+	ResponseTimeMs float64                   `json:"response_time_ms"`
 	Blocked        bool                      `json:"blocked"`
 	Cached         bool                      `json:"cached"`
 	BlockTrace     []storage.BlockTraceEntry `json:"block_trace,omitempty"`
